refactor(mrbgems): Relocate IO implementation to picoruby-machine
This commit refactors the IO implementation for non-POSIX environments
by moving it from `picoruby-io-console` and `picoruby-io` to
`picoruby-machine`.

The `gets` and `getc` methods, along with their C implementations,
have been moved from `picoruby-io-console` to `picoruby-machine`.

The `puts`, `print`, and `p` methods, previously defined in the Kernel
module, are now defined on the IO class within `picoruby-machine`.
The Ruby implementations of these methods have been moved from
`picoruby-io` to `picoruby-machine`.

This change centralizes the core IO functionalities for non-POSIX
systems into the `picoruby-machine` gem, improving modularity and
clarifying dependencies.
