+++ mrbgems/picoruby-io-console/include/io-console.h
@@ -6,9 +6,6 @@ extern "C" {
 #endif
 
 int hal_getchar(void);
-#if defined(PICORB_VM_MRUBY)
-int hal_write(int fd, const void *buf, int nbytes);
-#endif
 
 bool io_raw_q(void);
 void io_raw_bang(bool nonblock);
