+++ mrbgems/picoruby-machine/mrblib/0_io.rb
@@ -1,0 +1,7 @@
+class IO
+  if RUBY_ENGINE == "mruby"
+    def self.open(*args)
+      self.new # TODO
+    end
+  end
+end
