--- mrbgems/picoruby-machine/mrblib/machine.rb
@@ -1,2 +1,0 @@
-class Machine
-end
